* master: (51 commits)
  Fix fee estimation when processing invites
  Fix use of invalidated iterator which can cause a crash.
  fix parens position
  fix help message
  fixes for WSL bash
  signing script
  fix alignment on toolbar buttons to be consistent when selected
  allow network and transaction listviews to scroll
  Fix potential divide by zero in the progress dialog.
  Bumping version to 0.3.3
  Change progress per hour to blocks per hour.
  Fix computation of download estimate.
  Fix indexing transactions.
  sign package automagically if variables provided
  Signing script
  sign OSX dmg file
  script
  some compiler wornings fixed
  Bitcore -> Merit
  Merit supports 16MB blocks.
  ...

+++ src/consensus/merkle.cpp
@@ -188,7 +188,7 @@ uint256 BlockWitnessMerkleRoot(const CBlock& block, bool* mutated)
     }
 
     if(!block.invites.empty()) {
-        int s = 0;
+        long unsigned int s = 0;
         if(block.invites[0]->IsCoinBase()) {
             leaves[i].SetNull(); // The witness hash of the invite coinbase is 0.
             i++;
